-Edit UDI script D608 to be source specific,adapt to column names (CD,NAME,DESC) -Fixes in BMAP_CD,BMAP_CHECK testing scripts

+++ read_smx_sheet/templates/D608.py
@@ -25,23 +25,31 @@ import traceback
 #     f.close()
 
 
-def d608(cf, source_output_path, Core_tables, BMAP_values):
+def d608(cf, source_output_path,source_name,STG_tables, Core_tables, BMAP_values):
     file_name = funcs.get_file_name(__file__)
     f = funcs.WriteFile(source_output_path, file_name, "sql")
     try:
+        src_code_set_names = funcs.get_src_code_set_names(STG_tables,source_name)
         code_set_names= TransformDDL.get_code_set_names(BMAP_values)
-        for code_set in code_set_names:
-
-            tbl_pk = TransformDDL.get_trgt_pk(Core_tables, code_set)
-            columns = TransformDDL.get_lkp_tbl_Cols(Core_tables, code_set)
-            # Bmap_Vals=TransformDDL.get_bmap_values_for_codeset(BMAP_values,code_set)
-            for bmap_values_indx, bmap_values_row in BMAP_values[(BMAP_values['Code set name'] == code_set) & (BMAP_values['Layer'] == 'CORE')][['EDW code','Description']].drop_duplicates().iterrows():
-                del_st = "DELETE FROM " + cf.core_table + "." + code_set + " WHERE " + tbl_pk + " = '" + str(bmap_values_row['EDW code']) + "';\n"
-                insert_into_st = "INSERT INTO " + cf.core_table + "." + code_set + "(" + columns + ")\nVALUES "
-                insert_values = "(" + str(bmap_values_row["EDW code"]) + ", '" + str(bmap_values_row["Description"]) + "');\n\n"
-                insert_st = insert_into_st + insert_values
-                f.write(del_st)
-                f.write(insert_st)
+        for code_set_name in src_code_set_names:
+            for code_set in code_set_names:
+                if code_set_name == code_set:
+                    tbl_pk = TransformDDL.get_trgt_pk(Core_tables, code_set)
+                    columns = TransformDDL.get_lkp_tbl_Cols(Core_tables, code_set)
+                    flag_look_up = TransformDDL.get_is_look_up_flag(Core_tables,code_set)
+                    print(flag_look_up)
+                    if flag_look_up:
+                        for bmap_values_indx, bmap_values_row in BMAP_values[(BMAP_values['Code set name'] == code_set) & (BMAP_values['Layer'] == 'CORE')][['EDW code','Description']].drop_duplicates().iterrows():
+                            del_st = "DELETE FROM " + cf.core_table + "." + code_set + " WHERE " + tbl_pk + " = '" + str(bmap_values_row['EDW code']) + "';\n"
+                            insert_into_st = "INSERT INTO " + cf.core_table + "." + code_set + "(" + columns + ")\nVALUES "
+                            insert_values = ''
+                            if columns.count(',') == 1:
+                                insert_values = "(" + str(bmap_values_row["EDW code"]) + ", '" + str(bmap_values_row["Description"]) + "');\n\n"
+                            elif columns.count(',') == 2:
+                                insert_values = "(" + str(bmap_values_row["EDW code"]) + ", '" + str(bmap_values_row["Description"]) + ", '" + str(bmap_values_row["Description"])  + "');\n\n"
+                            insert_st = insert_into_st + insert_values
+                            f.write(del_st)
+                            f.write(insert_st)
     except:
         funcs.TemplateLogError(cf.output_path, source_output_path, file_name, traceback.format_exc()).log_error()
